refactor(components): now component more flexible
Jira issue: VEGA-1731
