+++ src/components/Shared/Article/ArticleOptionsDropdown/EditArticleModal/EditArticleModal.css
@@ -12,7 +12,6 @@
 }
 
 .EditArticleModal__text-field {
-  position: relative;
   margin-top: 0;
 }
