+++ src/components/Shared/Article/ArticleWrapper.css
@@ -3,7 +3,6 @@
 }
 
 .ArticleWrapper__row {
-  position: relative;
   margin-top: var(--space-2xs);
   grid-column-gap: var(--space-2xs);
 }
