--- src/components/Shared/ErrorMessage/ErrorMessage.css
@@ -1,4 +1,0 @@
-.ErrorMessage {
-  position: absolute;
-  top: var(--space-2xl);
-}
