--- src/components/Shared/ErrorMessage/cn-error-message.ts
@@ -1,3 +1,0 @@
-import bemCn from 'bem-cn';
-
-export const cnErrorMessage = bemCn('ErrorMessage');
