+++ src/components/Shared/ErrorMessage/cn-validation.ts
@@ -1,0 +1,3 @@
+import bemCn from 'bem-cn';
+
+export const cnValidation = bemCn('Validation');
